Crate test_sockopt

Structs

GetsockoptArguments 🔒

SetsockoptArguments 🔒

Functions

bufsize_test_helper 🔒

check_getsockopt_call 🔒

check_setsockopt_call 🔒

get_tests 🔒

main 🔒

test_invalid_fd 🔒

Test getsockopt() and setsockopt() using an argument that cannot be a fd.

test_invalid_level 🔒

Test getsockopt() and setsockopt() using an invalid level.

test_long_len 🔒

Test getsockopt() and setsockopt() using a non-null optval and a long optlen.

test_non_existent_fd 🔒

Test getsockopt() and setsockopt() using an argument that could be a fd, but is not.

test_non_socket_fd 🔒

Test getsockopt() and setsockopt() using a valid fd that is not a socket.

test_null_val 🔒

Test getsockopt() and setsockopt() using a null optval and a correct optlen.

test_null_val_nonzero_len 🔒

Test getsockopt() and setsockopt() using a null optval and a non-zero optlen.

test_null_val_null_len 🔒

Test getsockopt() using a null optval and a null optlen.

test_null_val_zero_len 🔒

Test getsockopt() and setsockopt() using a null optval and a zero optlen.

test_short_len 🔒

Test getsockopt() and setsockopt() using a non-null optval and a short optlen.

test_so_error 🔒

Test getsockopt() and setsockopt() using the SO_ERROR option.

test_so_rcvbuf 🔒

Test getsockopt() and setsockopt() using the SO_RCVBUF option.

test_so_sndbuf 🔒

Test getsockopt() and setsockopt() using the SO_SNDBUF option.

test_tcp_congestion 🔒

Test getsockopt() and setsockopt() using the TCP_CONGESTION option.

test_tcp_info 🔒

Test getsockopt() and setsockopt() using the TCP_INFO option.

test_tcp_nodelay 🔒

Test getsockopt() and setsockopt() using the TCP_NODELAY option.

test_zero_len 🔒

Test getsockopt() and setsockopt() using a non-null optval and a zero optlen.